chore: Keep column automations in the ESLint budget
The bindings and menu copy lived in one file that crossed the
max-lines budget. Split them so CI lint stays within the limit.
