+++ web_src/src/pages/factories/lib/columnAutomationsCopy.ts
@@ -1,0 +1,19 @@
+export const COLUMN_AUTOMATIONS_COPY = {
+  menuLabel: "Automations",
+  addLabel: "New automation",
+  addHint: "Create a canvas and open the editor.",
+  pickerTitle: "Add automation",
+  pickerDescription: "Choose an automation for this column.",
+  sourceTaken: "This automation is already configured.",
+  needsRepairLabel: "Needs repair",
+  disabledLabel: "Disabled",
+  editLabel: "Edit automation",
+  tabsLabel: "Automation sections",
+  generalTab: "General",
+  agentTab: "Agent",
+  automationTab: "Automation",
+  viewLoading: "The automation is loading.",
+  viewEmpty: "This automation has no canvas yet.",
+  viewError: "SuperPlane could not load the automation.",
+  viewRetry: "Try again",
+} as const;
